feat(e2e-utils): share local-registry helper across e2e tests
Moves the ph-registry lifecycle + auth helpers (startRegistry,
stopRegistry, createTestUser, writeNpmrc, verifyPublish) from
vetra-e2e's local helpers into @powerhousedao/e2e-utils so a second
suite (test/package-e2e) can consume them. Adds an optional
localPackages arg threaded through to the upcoming ph-registry
--local-packages flag. Vetra-e2e is migrated to the shared import.
